move MarkdownContext into MdxRenderer, since we need to get this into /client
