+++ packages/nextjs-contentful-renderer/src/MdxRenderer/MdxRenderer.tsx
@@ -15,3 +15,6 @@ export const MdxRenderer: FC<MdxRendererProps> = ({ content = {}, componentMap =
 
   return !isClient ? null : <MdxComponent components={{ ...components, ...componentMap }} />;
 };
+
+export * from './MarkdownContext';
+export * from './MarkdownContext/types';
