+++ packages/nextjs-contentful-renderer/src/MdxRenderer/renderers/link.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, LinkProps } from '@cmpsr/components';
-import { useMarkdown } from '../../MarkdownContext';
+import { useMarkdown } from '../MarkdownContext';
 
 export const renderLink = ({ children, ...props }: LinkProps) => {
   const { link = {} } = useMarkdown();
