+++ packages/nextjs-contentful-renderer/src/MdxRenderer/renderers/list.tsx
@@ -1,6 +1,6 @@
 import { ListItemProps, OrderedListProps, OrderedList, UnorderedList, UnorderedListProps } from '@cmpsr/components';
 import React from 'react';
-import { useMarkdown } from '../../MarkdownContext';
+import { useMarkdown } from '../MarkdownContext';
 
 export const renderOrderedList = (props: OrderedListProps) => {
   const { paragraph = {} } = useMarkdown();
