+++ packages/nextjs-contentful-renderer/src/MdxRenderer/renderers/text.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text, TextProps } from '@cmpsr/components';
-import { useMarkdown } from '../../MarkdownContext';
+import { useMarkdown } from '../MarkdownContext';
 
 export const renderText = ({ children, ...props }: TextProps, as: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p') => {
   const variants = {
