+++ packages/nextjs-contentful-renderer/src/index.ts
@@ -3,5 +3,3 @@ export * from './getPageContent';
 export * from './getDefaultTheme';
 export * from './getStaticPageContent';
 export * from './getStaticRoutes';
-export * from './MarkdownContext';
-export * from './MarkdownContext/types';
